Fixed misspellings
