+++ Client/__init__.py
@@ -2,18 +2,16 @@ import sys
 from PyQt5 import QtWidgets
 from PyQt5.QtGui import QPixmap
 from PyQt5.QtWidgets import QApplication, QMainWindow, QLabel, QLineEdit, QComboBox, QMessageBox
-from Clinet.OnePlayer import One
-from Clinet.MultiPlayer import Multi
+from Client.OnePlayer import One
+from Client.MultiPlayer import Multi
 from core.utils.image_helper import get_full_image_path
-
-
 
 
 class MyWindow(QMainWindow):
     def __init__(self):
         super(MyWindow, self).__init__()
         self.setGeometry(200, 200, 1000, 600)
-        self.setWindowTitle("Manu")
+        self.setWindowTitle("Menu")
         self.initUI()
 
     def initUI(self):
@@ -27,10 +25,10 @@ class MyWindow(QMainWindow):
         self.b1.setStyleSheet(
             "border:2px solid rgb(120, 20, 60); color: blue;font-size: 26px; font-family: Arial Black;");
         self.b1.clicked.connect(self.on_push_button)
-        self.dialog= OnePlayerWindow(self)
+        self.dialog = OnePlayerWindow(self)
 
         self.b2 = QtWidgets.QPushButton(self)
-        self.b2.setText("MULITPLAYER")
+        self.b2.setText("MULTIPLAYER")
         self.b2.setGeometry(400, 200, 250, 50)
         self.b2.setStyleSheet(
             "border:2px solid rgb(120, 20, 60); color: blue;font-size: 26px; font-family: Arial Black;");
@@ -52,6 +50,7 @@ class MyWindow(QMainWindow):
 
     def on_push_button(self):
         self.dialog.show()
+
     def on_push_button2(self):
         self.dialog2.show()
 
@@ -65,13 +64,16 @@ class MyWindow(QMainWindow):
         app = QApplication.instance()
         app.closeAllWindows()
 
+
 class OnePlayerWindow(One):
-    def __init__(self,parent=None):
+    def __init__(self, parent=None):
         super(OnePlayerWindow, self).__init__()
 
+
 class TwoPlayerWindow(Multi):
-    def __init__(self,parent=None):
+    def __init__(self, parent=None):
         super(TwoPlayerWindow, self).__init__()
+
 
 def window():
     app = QApplication(sys.argv)
@@ -81,4 +83,5 @@ def window():
     sys.exit(app.exec_())
 
 
-window()
+if __name__ == '__main__':
+    window()
